multi beta 1.1
added a function that will send mouse pos from server.py to client.py and will update client.py's mouse pos. client.py cannot be run standalone, but server.py will run alone standalone. so basically server.py needs to be running when client.py needs to be run.
with this we can exchange mouse position whenever it is one player's turn, but we cannot sync the data that both screens see. this means that both screens will end up having different questions. you cant emulate click with how the code is rn either. we'll have to find a way to sync the whole display as well as all inputs from one screen to the other and vice versa. or find a way for both players to run one client, which means dumping the idea that the server has the client stuff, then we update and move the mouse pos for client, which should sync mouse pos for both clients. then swap player turns for the question. hmm doable