New menu completed + qol features
Finished adding the history and books category questions. The picking mechanism for them is currently using random function. It'll have to be changed to manual picking after categories are added. Changed menu background theme to fit the atmosphere. Hovering mouse over menu options changes it to white, and plays a sfx. Clicking on the option plays another sfx.  Top text also changes when hovered over options. Added a bigger boxed background for when the question is longer. It now automatically adjusts and splits to fit in the pentagon. Added zen mode which has all the questions. Adjusted the choices menu to reflect on this. Also turned everything in the choices menu into smaller functions to reduce the code by a small amount.  Background music picking mechanic has been changed since we have too many questions for zen mode. Will have to be redone after categories are added.